Fix handling the end of JSON Arrays correctl (cf. jqassistant/jqa-json-plugin#11).
A base class (Listener pattern) offers several methods to handle
certain events while parsing the JSON data. There is a method to
handle the beginning and the end of an array, as well as methods to
handle the first and last elements of the array. All of them are empty
in the base class. This implementation overrides the methods to
enter/leave JSON arrays and JSON objects. A to deep nesting should be
prevented by counting the depth of nesting (I guess to prevent a stack
overflow?).

However, the current implementation overrides the handling of an array
start correctly but does not handle the end of the array
accordingly. It only works if the array has some content and a last
element event is issued by the parser. If the array is empty, no last
element exists and the handler for the last element is never called.
Hence, the nesting level is never decreased again in this case,
leading to the exception for very large JSONs with more than 999 empty
arrays.